requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
boto3>=1.33.0
botocore>=1.21.0
pydantic==2.5.1
bedrock-llm==0.1.8.5
jinja2==3.1.4
requests==2.31.0
vnstock3==3.1.0
python-dotenv==1.0.1